收获确认、收费、报表部分问题依照局方要求修改
1、收获确认输入小数无法保存问题
2、收费管理添加收费类型
3、报表添加根据单位搜索功能

+++ product/src/main/java/com/product/controller/BusinessControllerHelper.java
@@ -61,7 +61,7 @@ public class BusinessControllerHelper {
 		receiveinfo.setIrradtime(Common.stringDefaultOfEmpty(request.getParameter("irradtime"),""));
 		receiveinfo.setIrradtimeorg(Common.stringDefaultOfEmpty(request.getParameter("timeorgs"),""));
 		receiveinfo.setIrradflag(Common.stringDefaultOfEmpty(request.getParameter("irradflags"),""));
-		receiveinfo.setFee(Integer.valueOf(Common.stringDefaultOfEmpty(request.getParameter("fee"),"0")) );
+		receiveinfo.setFee(Double.valueOf(Common.stringDefaultOfEmpty(request.getParameter("fee"),"0")) );
 		receiveinfo.setMask(Common.stringDefaultOfEmpty(request.getParameter("mask"),""));
 		receiveinfo.setId(idAndreceiveorgid.split("@_@")[0]);
 		receiveinfo.setReceivemgrid(idAndreceiveorgid.split("@_@")[1]);
